+++ crates/fuel-core/src/query/message.rs
@@ -281,10 +281,13 @@ fn message_receipts_proof<T: MessageProofData + ?Sized>(
     }
 }
 
-pub fn message_status<T: DatabaseMessages + ?Sized>(
+pub fn message_status<T>(
     database: &T,
     message_nonce: Nonce,
-) -> StorageResult<MessageStatus> {
+) -> StorageResult<MessageStatus>
+where
+    T: OffChainDatabase + DatabaseMessages + ?Sized,
+{
     if database.message_is_spent(&message_nonce)? {
         Ok(MessageStatus::spent())
     } else if database.message_exists(&message_nonce)? {