+++ crates/fuel-core/src/service/genesis/importer.rs
@@ -9,23 +9,14 @@ use self::{
         Target,
     },
 };
-
 use super::task_manager::TaskManager;
-mod import_task;
-mod off_chain;
-mod on_chain;
-mod progress;
-use std::{
-    io::IsTerminal,
-    marker::PhantomData,
-};
-
 use crate::{
     combined_database::CombinedDatabase,
     database::database_description::{
         off_chain::OffChain,
         on_chain::OnChain,
     },
+    fuel_core_graphql_api::storage::messages::SpentMessages,
     graphql_api::storage::{
         coins::OwnedCoins,
         contracts::ContractsInfo,
@@ -41,6 +32,7 @@ use crate::{
         },
     },
 };
+use core::marker::PhantomData;
 use fuel_core_chain_config::{
     AsTable,
     SnapshotReader,
@@ -59,6 +51,7 @@ use fuel_core_storage::{
         ContractsState,
         FuelBlocks,
         Messages,
+        ProcessedTransactions,
         SealedBlockConsensus,
         Transactions,
     },
@@ -70,7 +63,15 @@ use fuel_core_types::{
     },
     fuel_types::BlockHeight,
 };
+use std::io::IsTerminal;
 use tracing::Level;
+
+mod import_task;
+mod off_chain;
+mod on_chain;
+mod progress;
+
+const GROUPS_NUMBER_FOR_PARALLELIZATION: usize = 10;
 
 pub struct SnapshotImporter {
     db: CombinedDatabase,
@@ -117,9 +118,11 @@ impl SnapshotImporter {
         self.spawn_worker_on_chain::<ContractsLatestUtxo>()?;
         self.spawn_worker_on_chain::<ContractsState>()?;
         self.spawn_worker_on_chain::<ContractsAssets>()?;
+        self.spawn_worker_on_chain::<ProcessedTransactions>()?;
 
         self.spawn_worker_off_chain::<TransactionStatuses, TransactionStatuses>()?;
         self.spawn_worker_off_chain::<OwnedTransactions, OwnedTransactions>()?;
+        self.spawn_worker_off_chain::<SpentMessages, SpentMessages>()?;
         self.spawn_worker_off_chain::<Messages, OwnedMessageIds>()?;
         self.spawn_worker_off_chain::<Coins, OwnedCoins>()?;
         self.spawn_worker_off_chain::<FuelBlocks, OldFuelBlocks>()?;
@@ -147,6 +150,10 @@ impl SnapshotImporter {
         let groups = self.snapshot_reader.read::<TableBeingWritten>()?;
         let num_groups = groups.len();
 
+        if num_groups == 0 {
+            return Ok(());
+        }
+
         let block_height = *self.genesis_block.header().height();
         let da_block_height = self.genesis_block.header().da_height;
         let db = self.db.on_chain().clone();
@@ -163,7 +170,14 @@ impl SnapshotImporter {
                 db,
                 progress_reporter,
             );
-            tokio_rayon::spawn(move || task.run())
+            async move {
+                if num_groups >= GROUPS_NUMBER_FOR_PARALLELIZATION {
+                    tokio_rayon::spawn(move || task.run()).await?;
+                } else {
+                    task.run()?;
+                }
+                Ok(())
+            }
         });
 
         Ok(())
@@ -182,6 +196,11 @@ impl SnapshotImporter {
     {
         let groups = self.snapshot_reader.read::<TableInSnapshot>()?;
         let num_groups = groups.len();
+
+        if num_groups == 0 {
+            return Ok(());
+        }
+
         let block_height = *self.genesis_block.header().height();
         let da_block_height = self.genesis_block.header().da_height;
 
@@ -200,7 +219,14 @@ impl SnapshotImporter {
                 db,
                 progress_reporter,
             );
-            tokio_rayon::spawn(move || task.run())
+            async move {
+                if num_groups >= GROUPS_NUMBER_FOR_PARALLELIZATION {
+                    tokio_rayon::spawn(move || task.run()).await?;
+                } else {
+                    task.run()?;
+                }
+                Ok(())
+            }
         });
 
         Ok(())